requirements.txt
aiogram==3.19.0
python-dotenv
yt-dlp
clck
